eototoeoto: 어드민 컨트롤러 수정

+++ src/main/java/NoTurningBack/jinddobey/repository/DealRepository.java
@@ -3,5 +3,5 @@ package NoTurningBack.jinddobey.repository;
 import NoTurningBack.jinddobey.domain.Deal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DealRepository extends JpaRepository<Deal, Integer> {
+public interface DealRepository extends JpaRepository<Deal, Long> {
 }
